requirements.txt
python-vlc==1.1.2
cycler==0.10.0
matplotlib==1.5.3
numpy==1.11.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
typeguard==2.1.1
